+++ src/lib.rs
@@ -44,14 +44,6 @@
 //!     world.run();
 //! }
 //! ```
-pub use self::entity::{
-    Component, ComponentBox, Entity, EntityBuilder, EntityComponentManager, EntityContainer,
-    SharedComponentBox, VecEntityContainer,
-};
-pub use self::error::NotFound;
-pub use self::system::{EntitySystem, EntitySystemBuilder, EntitySystemManager, Priority, System};
-pub use self::world::World;
-
 pub mod entity;
 pub mod error;
 pub mod prelude;