+++ src/prelude.rs
@@ -1,4 +1,5 @@
-pub use {
-    Component, ComponentBox, Entity, EntityComponentManager, EntityContainer, Priority, System,
-    VecEntityContainer, World,
+pub use crate::{
+    entity::{Component, ComponentBox, Entity, EntityComponentManager, VecEntityContainer},
+    system::{Priority, System},
+    world::World,
 };